Move BOM-related methods to IOUtils where other BOM-methods live
