+++ src/org/opensolaris/opengrok/analysis/AnalyzerGuru.java
@@ -101,6 +101,7 @@ import org.opensolaris.opengrok.history.HistoryGuru;
 import org.opensolaris.opengrok.history.HistoryReader;
 import org.opensolaris.opengrok.logger.LoggerFactory;
 import org.opensolaris.opengrok.search.QueryBuilder;
+import org.opensolaris.opengrok.util.IOUtils;
 import org.opensolaris.opengrok.web.Util;
 
 /**
@@ -851,12 +852,12 @@ public class AnalyzerGuru {
 
         in.mark(512);
 
-        String encoding = findBOMEncoding(sig);
+        String encoding = IOUtils.findBOMEncoding(sig);
         if (encoding == null) {
             encoding = "UTF-8";
         } else {
-            byte[] bom = BOMS.get(encoding);
-            if (in.skip(bom.length) < bom.length) {
+            int skipForBOM = IOUtils.skipForBOM(sig);
+            if (in.skip(skipForBOM) < skipForBOM) {
                 in.reset();
                 return "";
             }
@@ -904,56 +905,5 @@ public class AnalyzerGuru {
 
         in.reset();
         return opening.toString();
-    }
-
-    /**
-     * Byte-order markers.
-     */
-    private static final Map<String, byte[]> BOMS
-            = new HashMap<>();
-
-    static {
-        BOMS.put("UTF-8", new byte[]{(byte) 0xEF, (byte) 0xBB, (byte) 0xBF});
-        BOMS.put("UTF-16BE", new byte[]{(byte) 0xFE, (byte) 0xFF});
-        BOMS.put("UTF-16LE", new byte[]{(byte) 0xFF, (byte) 0xFE});
-    }
-
-    /**
-     * Gets a value indicating a UTF encoding if the array starts with a
-     * known byte sequence
-     *
-     * @param sig a sequence of bytes to inspect for a BOM
-     * @return null if no BOM was identified; otherwise a defined charset name
-     */
-    private static String findBOMEncoding(byte[] sig) {
-        for (Map.Entry<String, byte[]> entry : BOMS.entrySet()) {
-            String encoding = entry.getKey();
-            byte[] bom = entry.getValue();
-            if (sig.length > bom.length) {
-                int i = 0;
-                while (i < bom.length && sig[i] == bom[i]) {
-                    i++;
-                }
-                if (i == bom.length) return encoding;
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Gets a value indicating the number of UTF BOM bytes at the start of an
-     * array
-     *
-     * @param sig a sequence of bytes to inspect for a BOM
-     * @return 0 if the array doesn't start with a BOM; otherwise the number of
-     * BOM bytes
-     */
-    public static int skipForBOM(byte[] sig) {
-        String encoding = findBOMEncoding(sig);
-        if (encoding != null) {
-            byte[] bom = BOMS.get(encoding);
-            return bom.length;
-        }
-        return 0;
     }
 }
